Parse & stringify gql when documentMode=string
documentMode=string is perfect for graphql-request, but it simply
inlines any referenced fragments. This means that fragments that
are included more than once in a single query for example, by being
included by another fragment in that query, as well as at the top level -
are thus duplicated in the query, and then fail when send to the server.

This PR parses and stringifies queries ahead of time such that
no duplicates occur.