Include undefined as type for all-optional argument lists
The motivation behind this change is lined out in more detail in
dotansimha/graphql-code-generator#3322. The basic problem is that
`RequireFields` includes provisions to make non-required arguments
optional (via `?:` mapping) but this wrapper type only gets applied
when there is at least one required argument in the list.

However, lists with all-optional arguments must have the optional
mapping applied as well. If not, they end up including only `null` but
not `undefined` in their types which can lead to subtle runtime errors.