feat: allow field definitions to be wrapped (#3299)
* feat: allow field definitions to be wrapped

Adds the optional (disabled by default) `FieldWrapper` type for wrapping
field definitions on objects and interfaces (but _not_ input types).

In addition to returning a value of the actual type for a field, objects
can also return Promises to that type or functions that do the same. It
is currently not possible to achieve that by reconfiguring the code
generator.

This adds the ability to wrap field definitions in a `FieldWrapper`,
which defaults to `FieldWrapper<T> = T`, thus not changing behavior in
any way. To get the above-described behavior, a user would need to set
`fieldWrapperValue` to `T | Promise<T> | (() => T | Promise<T>)` or
something similar in addition to enabling field wrapping by setting
`wrapFieldDefinitions` to `true`.

That may be overkill, but I wanted to avoid altering the current
existing behavior unexpectedly.

Closes: #3112

* feat: add wrapFieldDefinitions support to ts-resolvers