docs: name example query in React guide (#7817)
In order to be able to import the following:

```typescript
import { postsQueryDocument } from "./graphql/generated"
```

one needs to name the query in the `.graphql` file accordingly. Following the guide (and therefore not doing so) can lead to confusion as the generated document will just be called `Document` instead of `postsQueryDocument`, leading to various issues and headaches.