feat: Add option to squash exactly similar fragment types (#6826)
* feat: Add option to squash exactly similar fragment types

When selecting from an interface, the default code path will declare the
fragment type as a union of all implementing types. This happens
regardless of whether we actually selecting something from the
implementing types.

E.g.
```
type FooFragment =
  | {
    __typename: "foo";
    shared: string;
  }
  | {
    __typename: "bar";
    shared: string;
  }
```

This behavior can be noisy and redundant. In combination with heavy use of
interfaces, especially in nested schemas, this can also potentially
trigger TypeScript's "union type is too complex to represent".

In order to reduce the size of the generated fragments, we introduce the
`mergeFragmentTypes` configuration option. When enabled, this will
collapse all types with the same selection set to a single union member
of the fragment type

E.g.
```
type FooFragment = {
  __typename: "foo" | "bar";
  shared: string;
}
```

* Support skipTypename

* Avoid mangling fragment type names unless we have to

* Work around TS limits

* Fix tests

* Disable interface merging when masking inline fragments.

While we'd ideally make this work, the way we determine type uniqueness
would probably require a two-pass approach,n since we need to determine
whether a type can be compacted before we add the fragment reference to
the type (just adding the fragment reference straight away makes all
types unique).