feat: add option to add `__typename` to interfaces
The new option `addTypenameToInterfaces` adds the `__typename` field to
types resulting from interfaces. The field admits a union of string
values, each the name of one of the concrete types implementing the
interface.

Also adds field `typeUnionOperator` to class `BaseVisitor` so we don't
have the literal `' | '` in a bunch of places.